Fix dev server autoreload interaction with streams
Previously, one stream process would be started each time the
autoreloader restarted the dev server. Every one but the first would
immediately crash due to not being able to bind the correct port.

Now we close down the stream server (relatively) gracefully on exit, and
start a new one on the autoreloader.

As a nice bonus, there's no longer an arbitrary delay between starting the dev
server and the stream server coming up.