Prefer Python 2
There's a bug in Python 3 which means we don't want to be using it at
the competition, yet this is probably the best general installation
script we have for the system so it's likely to be what we use to
deploy onto the casual machines (ie the scorers and anyone else who
needs to poke things) at the competition. Therefore it should prefer
the more functional Python version.

Change-Id: Ia290d9be011b83c16e14af536004c558d13a028e